Combat audiosystem
When a fight occurs a combat song will start playing. It works by fading
out the looping song, and when the combat is over there is a timer that
will go from 0 to 1 before the code starts to fade out the combat song
and fade in the background song. This is to prevent the code from
stopping the combat song when there might be a new ship ready for fight
the very instance the other ship died.