New sound system for when to start and stop combat songs
We changed how we wanted combat music to work, so I had to change the
code a little. We have 1 combat song for each looping background song,
so the songs have to be in sync.
I pretty much just made it so both sound sources always plays, and only
the combat source volume is modified.
When the code changes background song, it also changes combat song.

This works pretty well.