First iteration of player meeting boss
If the player drives up to the boss, he/she will see that the boss is
stationary. When close enough, the players ship will stop, and the boss
will rotate. It will stop rotating when facing the player. The thought
is to start the dialogue when the bool "faced" == true