Boss behaviour when player has died or fled from combat, and first encounter improvement.
Boss will fight after the first dialogue with it is over, and the
dialogue will only occur once every play session. Bombs despawn when
player dies and when the boss is despawned. The boss and bombs despawn
when player is far enough away from the player, just like other ai's
despawn.