Renamed Cursor script, added aim cursor when mouse2 is down
When the player aim (right click) the aim cursor will appear. It will
return to the default one when mouse2 is no longer pressed.
I had to rename the Cursor script to something else, as Unity has a
built in functionality called "Cursor", so the script we made conflicted
with this built in functionality.