New combat song, updated economy, can no longer zoom while in worldmap
I added the new combat song made by Joakim, so now we are almost there!
I also updated the economy. It is in no way balanced yet, but it is
better than before. The new plan is to make a scraps value determined by
the cannonlevels on the ship they are spawning from.

It is no longer available to zoom while the big map is open. The main
camera will zoom out to the max zoom, and return to the playerchosen
zoom afterwards.