AI Patroling: No longer spawns gameobject
It used to be so that whenever an ai spawned, it would spawn with a
gameobject it would patrole towards. This was bad. It required the
computer not just to spawn the ai, but also an extra gameobject for the
sake of getting some Vector3 positions.

So now the code just generates a vector3 coordinate, and the ai will
drive towards this one instead of the gameobject. Woho!