Fixed rare case where collisions would return null in array
In ShipHitObjects the script takes the object it hit and dealt damage to
it. Sometimes a bug occurs and the return value is null. The code now
checks if it is null, and if it isn't then we can continue.

This should fix the issue "Being in combat while there are noe living
ai's".