"Aiming" icon now placed correctly when player is aiming
Instead of changing from one cursor to another when the player aims with
the special weapon, I had to create a RawImage component in the Canvas.
The image it uses is that of the "Aimed Gun" variety (wat), and it's
position is relevant to that of the cursor.

When the player aims, aka right clicks, the RawImage is visible, and the
cursor is invisible. The other way around when not right clicking.